fastapi_builder/templates/project/{{ cookiecutter.folder_name }}/requirements.txt
aiomysql==0.2.0
alembic==1.13.2
bcrypt==4.2.0
databases==0.9.0
email-validator==2.2.0
fastapi==0.112.2
loguru==0.7.2
passlib==1.7.4
pydantic==2.8.2
pymysql==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
sqlalchemy==2.0.32
uvicorn==0.30.6
